change(tests): all beacon root tests in the same file
